+++ src/main/java/edu/kit/datamanager/metastore2/Application.java
@@ -184,6 +184,8 @@ public class Application {
     IAuditService<ContentInformation> contentAuditService;
     MetastoreConfiguration rbc = new MetastoreConfiguration();
     rbc.setBasepath(this.applicationProperties.getSchemaFolder());
+    rbc.setAuthEnabled(this.applicationProperties.isAuthEnabled());
+    rbc.setJwtSecret(this.applicationProperties.getJwtSecret());
     rbc.setReadOnly(false);
     rbc.setDataResourceService(schemaResourceService);
     rbc.setContentInformationService(schemaInformationService);
@@ -205,6 +207,7 @@ public class Application {
     schemaResourceService.configure(rbc);
     schemaInformationService.configure(rbc);
     rbc.setAuditService(auditServiceDataResource);
+    rbc.setMaxJaversScope(this.applicationProperties.getMaxJaversScope());
     rbc.setSchemaRegistries(checkRegistries(applicationProperties.getSchemaRegistries()));
     rbc.setValidators(validators);
     MetadataRecordUtil.setSchemaConfig(rbc);
@@ -225,6 +228,8 @@ public class Application {
     IAuditService<ContentInformation> contentAuditService;
     MetastoreConfiguration rbc = new MetastoreConfiguration();
     rbc.setBasepath(applicationProperties.getMetadataFolder());
+    rbc.setAuthEnabled(this.applicationProperties.isAuthEnabled());
+    rbc.setJwtSecret(this.applicationProperties.getJwtSecret());
     rbc.setReadOnly(false);
     rbc.setDataResourceService(dataResourceService);
     rbc.setContentInformationService(contentInformationService);
@@ -246,6 +251,7 @@ public class Application {
     dataResourceService.configure(rbc);
     contentInformationService.configure(rbc);
     rbc.setAuditService(auditServiceDataResource);
+    rbc.setMaxJaversScope(this.applicationProperties.getMaxJaversScope());
     rbc.setSchemaRegistries(checkRegistries(applicationProperties.getSchemaRegistries()));
     rbc.setValidators(validators);