+++ src/main/java/edu/kit/datamanager/metastore2/configuration/ApplicationProperties.java
@@ -52,4 +52,7 @@ public class ApplicationProperties extends GenericApplicationProperties{
   @Value("${metastore.metadata.schemaRegistries: }")
   private String[] schemaRegistries;
 
+  @Value("${metastore.javers.scope:20}")
+  private int maxJaversScope;
+
 }